backend/python_services/data_fetcher/requirements.txt
firebase-admin>=6.5.0
chromadb>=0.5.0
requests>=2.31.0
newsapi-python>=0.2.7
firecrawl-py>=0.0.13
jinaai>=0.2.1
google-generativeai>=0.5.2
python-dotenv>=1.0.1